[CLIENT] CategoryApiSlice code refactor
