+++ client/src/features/category/questions/QuestionList.js
@@ -5,7 +5,7 @@ import Header from '../../../components/Header'
 import Bluebutton from '../../../components/Bluebutton'
 import QuestionAdd from './QuestionAdd'
 import { useParams } from 'react-router'
-import { useGetQuestionsBySectionQuery } from '../categoryApiSlice'
+import { useGetQuestionsBySectionQuery } from './questionApiSlice'
 import Question from './Question'
 import { Button, Table } from 'flowbite-react'
 import { TableHead } from 'flowbite-react/lib/esm/components/Table/TableHead'
